add two functons for both realstream and static
