+++ mdfrWeb/templates/index.html
@@ -50,11 +50,15 @@
         <div class="row">
             <div class="col-sm-6">
                 <h3>Realtime Stream</h3>
-                <img src="{{ url_for('static', filename='img/video.png')}}" class="img-responsive" style="width:100%" alt="Image">
+                <a href="/realStream">
+                    <img src="{{ url_for('static', filename='img/video.png')}}" class="img-responsive" style="width:100%" alt="Image">
+                </a>
             </div>
             <div class="col-sm-6">
                 <h3>Static Image/Video</h3>
-                <img src="{{ url_for('static', filename='img/image.png')}}" class="img-responsive" style="width:100%" alt="Image">
+                <a href="/staticStream">
+                    <img src="{{ url_for('static', filename='img/image.png')}}" class="img-responsive" style="width:100%" alt="Image">
+                </a>
             </div>
         </div>
     </div><br>
