+++ mdfrWeb/templates/realStream.html
@@ -16,7 +16,7 @@
 <body>
 
     <div id="faceShow">
-        <!--<img src="{{ url_for('video_feed') }}"> -->
+        <img src="{{ url_for('video_feed') }}">
     </div>
 
 <footer class="container-fluid text-center">
